// app/src/js/aurora/lib/types.ts
export type Color =
  | 'white'
  | 'transparent-white'
  | 'black'
  | 'blue'
  | 'secondary-blue'
  | 'transparent-blue'
  | 'red'
  | 'secondary-red'
  | 'transparent-red'
  | 'green'
  | 'secondary-green'
  | 'transparent-green'
  | 'yellow'
  | 'secondary-yellow'
  | 'transparent-yellow'

export type Size = 'xs' | 'sm' | 'md' | 'lg' | 'xl'

export type IconType = 'line' | 'fill'